+++ app/store/chat.ts
@@ -28,6 +28,7 @@ import {
 } from "../utils/prodinfo";
 import { useRAGStore } from "./rag";
 import { useKGStore } from "./kg";
+import { getVectorStoreServerGlobal } from "../utils/rag";
 
 const generateUniqId = () => uuidv4();
 
@@ -168,10 +169,37 @@ export const useChatStore = createPersistStore(
         }));
       },
 
+      initializeChat(mask?: Mask) {
+        if (!mask) {
+          return;
+        }
+        const sessions = _get().sessions;
+        if (sessions.length === 1 && sessions[0].messages.length === 0) {
+          const config = useAppConfig.getState();
+          this.updateCurrentSession((session: ChatSession) => {
+            session.topic = mask.name;
+            session.mask = {
+              ...mask,
+              modelConfig: {
+                ...config.modelConfig,
+                ...mask.modelConfig,
+              }
+            }
+          });
+        }
+      },
       selectSession(index: number) {
         set({
           currentSessionIndex: index,
         });
+
+        // on session changed
+        if (index >= get().sessions.length) {
+          return;
+        }
+        const model = get().sessions[index].mask?.modelConfig?.model ?? "gpt-3.5-turbo";
+        const sessionId = get().sessions[index].id;
+        useAccessStore.getState().updateTokenUsage(sessionId, model);
       },
 
       moveSession(from: number, to: number) {
@@ -339,10 +367,16 @@ export const useChatStore = createPersistStore(
         const vsInfo = getVectorStoreInfo(prodInfo);
         const kgInfo = getKnowledgeGraphInfo(prodInfo);
         const ragConfig = vsInfo.enabled ? useRAGStore.getState().currentRAGConfig() : undefined;
+        const globalVS = ragConfig !== undefined ? 
+          getVectorStoreServerGlobal(ragConfig.connectionArgs, prodInfo?.VectorStore?.servers??[]) :
+          false;
+        if (ragConfig && globalVS) {
+          ragConfig.docIdsWorkspace = undefined;
+        }  
         const useRAG = useChatStore.getState().currentSession().useRAGSession;
         const useKG = useChatStore.getState().currentSession().useKGSession;
         const useOncoKB = useChatStore.getState().currentSession().useOncoKBSession??false;
-        const useAutoAgent = useChatStore.getState().currentSession().useAutoAgentSession;
+        const useAutoAgent = false; // useChatStore.getState().currentSession().useAutoAgentSession;
         const kgConfig = kgInfo.enabled ? useKGStore.getState().config : undefined;
         const oncokbConfig = oncokbInfo.enabled ? {
           useOncoKB: useOncoKB,
@@ -381,6 +415,7 @@ export const useChatStore = createPersistStore(
               get().onNewMessage(botMessage);
             }
             ChatControllerPool.remove(session.id, botMessage.id);
+            useAccessStore.getState().updateTokenUsage(session.id, modelConfig.model);
           },
           onError(error) {
             const isAborted = error.message.includes("aborted");
@@ -579,7 +614,7 @@ export const useChatStore = createPersistStore(
 
         const historyMsgLength = countMessages(toBeSummarizedMsgs);
 
-        if (historyMsgLength > modelConfig?.max_tokens ?? 4000) {
+        if (historyMsgLength > (modelConfig?.max_tokens ?? 4000)) {
           const n = toBeSummarizedMsgs.length;
           toBeSummarizedMsgs = toBeSummarizedMsgs.slice(
             Math.max(0, n - modelConfig.historyMessageCount),