+++ app/utils/rag.ts
@@ -9,7 +9,7 @@ const getConnectionArgsToDisplay = (
     if (server.server === connectionArgs.host) {
       return {
         host: server.server,
-        port: `${server.port}` ?? defaultPort,
+        port: `${server.port ?? defaultPort}`,
       }
     }
     const serverPort = `${server.port??defaultPort}`;
@@ -18,7 +18,7 @@ const getConnectionArgsToDisplay = (
       && (serverPort === connectionArgsPort) ) {
       return {
         host: server.server,
-        port: server.port ?? defaultPort,
+        port: serverPort,
       }
     }
   }
@@ -37,6 +37,24 @@ export const getVectorStoreConnectionArgsToDisplay = (
   vsServers: Array<DbServerSettings>,
 ): DbConnectionArgs => {
   return getConnectionArgsToDisplay(connectionArgs, vsServers, "19530");
+}
+export const getVectorStoreServerGlobal = (
+  connectionArgs: DbConnectionArgs,
+  servers: Array<DbServerSettings>,
+): boolean => {
+  const defaultPort = "19530";
+  for (const server of servers) {
+    if (server.server === connectionArgs.host) {
+      return server.global ?? false;
+    }
+    const serverPort = `${server.port??defaultPort}`;
+    const connectionArgsPort = `${connectionArgs.port??defaultPort}`
+    if (server.address === connectionArgs.host 
+      && (serverPort === connectionArgsPort) ) {
+      return server.global ?? false;
+    }
+  }
+  return false;
 }
 
 const getConnectionArgsToConnect = (