+++ CHANGELOG.md
@@ -2,6 +2,45 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.4.0](https://github.com/fengsh27/chatgse-next-sync/compare/v1.3.5...v1.4.0) (2025-01-06)
+
+
+### Features
+
+* display token usage ([#14](https://github.com/fengsh27/chatgse-next-sync/issues/14)) ([a9565a4](https://github.com/fengsh27/chatgse-next-sync/commit/a9565a449b852123728651a8bbe816e664ec9668))
+
+### [1.3.5](https://github.com/biocypher/biochatter-next/compare/v1.3.4...v1.3.5) (2024-08-26)
+
+
+### Bug Fixes
+
+* bugs in store initalization ([#75](https://github.com/biocypher/biochatter-next/issues/75)) ([ff2bb1b](https://github.com/biocypher/biochatter-next/commit/ff2bb1b4a7ee949543ce29e71dabe1107dbae821))
+
+### [1.3.4](https://github.com/biocypher/biochatter-next/compare/v1.3.3...v1.3.4) (2024-08-24)
+
+
+### Bug Fixes
+
+* rag store initialization bug and remove other gpt models ([#74](https://github.com/biocypher/biochatter-next/issues/74)) ([a6e9a65](https://github.com/biocypher/biochatter-next/commit/a6e9a65ec3bdaa2a09789a3bce05aba439456d37))
+
+### [1.3.3](https://github.com/biocypher/biochatter-next/compare/v1.3.2...v1.3.3) (2024-08-23)
+
+
+### Bug Fixes
+
+* 'New Conversation' issue ([#73](https://github.com/biocypher/biochatter-next/issues/73)) ([ad4952d](https://github.com/biocypher/biochatter-next/commit/ad4952d13bd1f3e1c324bb59b6feab0046d41ad9))
+
+### [1.3.2](https://github.com/biocypher/biochatter-next/compare/v1.3.1...v1.3.2) (2024-08-21)
+
+### [1.3.1](https://github.com/biocypher/biochatter-next/compare/v1.3.0...v1.3.1) (2024-08-19)
+
+## [1.3.0](https://github.com/biocypher/biochatter-next/compare/v1.2.0...v1.3.0) (2024-08-19)
+
+
+### Features
+
+* support OncoKB api agent and auto agent selector ([#70](https://github.com/biocypher/biochatter-next/issues/70)) ([6ff977c](https://github.com/biocypher/biochatter-next/commit/6ff977c657c6b51dc6211c72b56ae5bcb44b2363))
+
 ## [1.2.0](https://github.com/biocypher/biochatter-next/compare/v1.1.0...v1.2.0) (2024-07-05)